+++ CHANGELOG.md
@@ -8,6 +8,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.10.0] - 2025-11-19
+
+* [PR-246](https://github.com/itk-dev/economics/pull/246)
+  Minor api adjustments.
+* [PR-243](https://github.com/itk-dev/economics/pull/243)
+  Added CRUD for service agreements and cyber security agreements.
+  Added JSON endpoint for retrieving agreements.
+* [PR-240](https://github.com/itk-dev/economics/pull/240)
+  Fixes and optimizations for app:sync command
 * [PR-235](https://github.com/itk-dev/economics/pull/235)
   Increased timeout of nginx.
   Included archived issues in sync.
@@ -500,7 +509,8 @@ complete process.
 * Updated to authorization code flow.
 * Changed worklog save button styling to be sticky.
 
-[Unreleased]: https://github.com/itk-dev/economics/compare/2.9.4...HEAD
+[Unreleased]: https://github.com/itk-dev/economics/compare/2.10.0...HEAD
+[2.10.0]: https://github.com/itk-dev/economics/compare/2.9.4...2.10.0
 [2.9.4]: https://github.com/itk-dev/economics/compare/2.9.3...2.9.4
 [2.9.3]: https://github.com/itk-dev/economics/compare/2.9.2...2.9.3
 [2.9.2]: https://github.com/itk-dev/economics/compare/2.9.0...2.9.2